update Order and Item entities

+++ src/main/resources/db/migration/V2__CREATE_PRODUCT_TB.sql
@@ -1,6 +1,6 @@
 CREATE TABLE product
 (
-    id          INT NOT NULL,
+    id          BIGINT NOT NULL,
     description VARCHAR(255),
     price       DECIMAL,
     CONSTRAINT pk_product PRIMARY KEY (id)
